+++ mpitests/test_apply.py
@@ -16,6 +16,7 @@ Unit tests for apply on :mod:`pennylane_lightning` MPI-enabled devices.
 """
 # pylint: disable=protected-access,cell-var-from-loop,c-extension-no-member
 import itertools
+from functools import partial
 
 import numpy as np
 import pennylane as qml
@@ -652,9 +653,10 @@ class TestGenerateSample:
         """
         num_wires = numQubits
 
-        dev_mpi = qml.device(device_name, wires=num_wires, mpi=True, shots=2000, c_dtype=c_dtype)
+        dev_mpi = qml.device(device_name, wires=num_wires, mpi=True, c_dtype=c_dtype)
         dev_mpi._statevector.reset_state()
 
+        @partial(qml.set_shots, shots=2000)
         @qml.qnode(dev_mpi)
         def circuit():
             qml.RX(1.5708, wires=0)
@@ -671,8 +673,9 @@ class TestGenerateSample:
         """
         num_wires = 3
 
-        dev_mpi = qml.device(device_name, wires=num_wires, mpi=True, shots=2000, c_dtype=c_dtype)
+        dev_mpi = qml.device(device_name, wires=num_wires, mpi=True, c_dtype=c_dtype)
 
+        @partial(qml.set_shots, shots=2000)
         @qml.qnode(dev_mpi)
         def circuit():
             qml.Hadamard(0)
@@ -688,14 +691,13 @@ class TestGenerateSample:
         """Test that a tensor product involving PauliX and PauliY works correctly"""
         num_wires = 3
 
-        dev_mpi = qml.device(
-            device_name, wires=num_wires, mpi=True, shots=2000, c_dtype=c_dtype, seed=42
-        )
+        dev_mpi = qml.device(device_name, wires=num_wires, mpi=True, c_dtype=c_dtype)
 
         theta = 0.432
         phi = 0.123
         varphi = -0.543
 
+        @partial(qml.set_shots, shots=2000)
         @qml.qnode(dev_mpi)
         def circuit():
             qml.RX(theta, wires=[0])
@@ -730,14 +732,13 @@ class TestGenerateSample:
         """Test that a tensor product involving PauliZ and PauliY and hadamard works correctly"""
         num_wires = 3
 
-        dev_mpi = qml.device(
-            device_name, wires=num_wires, mpi=True, shots=2000, c_dtype=c_dtype, seed=42
-        )
+        dev_mpi = qml.device(device_name, wires=num_wires, mpi=True, c_dtype=c_dtype)
 
         theta = 0.432
         phi = 0.123
         varphi = -0.543
 
+        @partial(qml.set_shots, shots=2000)
         @qml.qnode(dev_mpi)
         def circuit():
             qml.RX(theta, wires=[0])
@@ -776,14 +777,13 @@ class TestTensorVar:
         """Test that a tensor product involving PauliX and PauliY works correctly"""
         num_wires = 3
 
-        dev_mpi = qml.device(
-            device_name, wires=num_wires, mpi=True, shots=2000, c_dtype=c_dtype, seed=42
-        )
+        dev_mpi = qml.device(device_name, wires=num_wires, mpi=True, c_dtype=c_dtype)
 
         theta = 0.432
         phi = 0.123
         varphi = -0.543
 
+        @partial(qml.set_shots, shots=2000)
         @qml.qnode(dev_mpi)
         def circuit():
             qml.RX(theta, wires=[0])
@@ -809,14 +809,13 @@ class TestTensorVar:
     def test_pauliz_hadamard(self, c_dtype, tol=TOL_STOCHASTIC):
         """Test that a tensor product involving PauliZ and PauliY and hadamard works correctly"""
         num_wires = 3
-        dev_mpi = qml.device(
-            device_name, wires=num_wires, mpi=True, shots=2000, c_dtype=c_dtype, seed=42
-        )
+        dev_mpi = qml.device(device_name, wires=num_wires, mpi=True, c_dtype=c_dtype)
 
         theta = 0.432
         phi = 0.123
         varphi = -0.543
 
+        @partial(qml.set_shots, shots=2000)
         @qml.qnode(dev_mpi)
         def circuit():
             qml.RX(theta, wires=[0])