+++ .github/CHANGELOG.md
@@ -52,6 +52,9 @@
   [(#1175)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1175)
 
 <h3>Internal changes ⚙️</h3>
+
+- Update tests with `device(..., shots=...)` to use `qml.set_shots` to ensure compatibility with the latest version of PennyLane.
+  [(#1173)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1173)
   
 - Fix `applyMultiRZ` for `lightning.kokkos` device to use Kokkos function instead of lambda.
   [(#1194)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1194)